// .PRD/erp-refactor/v3/types/sales-order-item.ts
import { z } from 'zod';

export const SalesOrderItemSchema = z.object({
  id: z.string(),
  owner: z.string().optional(),
  creation: z.string().optional(),
  modified: z.string().optional(),
  modified_by: z.string().optional(),
  fg_item: z.string().optional(),
  fg_item_qty: z.number().optional(),
  item_code: z.string(),
  customer_item_code: z.string().optional(),
  ensure_delivery_based_on_produced_serial_no: z.boolean().optional().default(false),
  is_stock_item: z.boolean().optional().default(false),
  reserve_stock: z.boolean().optional().default(true),
  delivery_date: z.string().optional(),
  item_name: z.string(),
  description: z.string().optional(),
  item_group: z.string().optional(),
  brand: z.string().optional(),
  image: z.string().optional(),
  image_view: z.string().optional(),
  qty: z.number(),
  stock_uom: z.string().optional(),
  subcontracted_qty: z.number().optional().default(0),
  uom: z.string(),
  conversion_factor: z.number(),
  stock_qty: z.number().optional(),
  stock_reserved_qty: z.number().optional().default(0),
  price_list_rate: z.number().optional(),
  base_price_list_rate: z.number().optional(),
  margin_type: z.enum(['Percentage', 'Amount']).optional(),
  margin_rate_or_amount: z.number().optional(),
  rate_with_margin: z.number().optional(),
  discount_percentage: z.number().optional(),
  discount_amount: z.number().optional(),
  distributed_discount_amount: z.number().optional(),
  base_rate_with_margin: z.number().optional(),
  rate: z.number().optional(),
  amount: z.number().optional(),
  item_tax_template: z.string().optional(),
  base_rate: z.number().optional(),
  base_amount: z.number().optional(),
  pricing_rules: z.string().optional(),
  stock_uom_rate: z.number().optional(),
  is_free_item: z.boolean().optional().default(false),
  grant_commission: z.boolean().optional().default(false),
  net_rate: z.number().optional(),
  net_amount: z.number().optional(),
  base_net_rate: z.number().optional(),
  base_net_amount: z.number().optional(),
  billed_amt: z.number().optional(),
  valuation_rate: z.number().optional(),
  gross_profit: z.number().optional(),
  delivered_by_supplier: z.boolean().optional().default(false),
  supplier: z.string().optional(),
  weight_per_unit: z.number().optional(),
  total_weight: z.number().optional(),
  weight_uom: z.string().optional(),
  warehouse: z.string().optional(),
  target_warehouse: z.string().optional(),
  prevdoc_docname: z.string().optional(),
  quotation_item: z.string().optional(),
  against_blanket_order: z.boolean().optional().default(false),
  blanket_order: z.string().optional(),
  blanket_order_rate: z.number().optional(),
  actual_qty: z.number().optional(),
  company_total_stock: z.number().optional(),
  bom_no: z.string().optional(),
  projected_qty: z.number().optional(),
  ordered_qty: z.number().optional(),
  planned_qty: z.number().optional(),
  production_plan_qty: z.number().optional(),
  work_order_qty: z.number().optional(),
  delivered_qty: z.number().optional(),
  produced_qty: z.number().optional(),
  returned_qty: z.number().optional(),
  picked_qty: z.number().optional(),
  additional_notes: z.string().optional(),
  page_break: z.boolean().optional().default(false),
  item_tax_rate: z.string().optional(),
  transaction_date: z.string().optional(),
  material_request: z.string().optional(),
  purchase_order: z.string().optional(),
  material_request_item: z.string().optional(),
  purchase_order_item: z.string().optional(),
  cost_center: z.string().optional().default(':Company'),
  project: z.string().optional(),
});

export type SalesOrderItem = z.infer<typeof SalesOrderItemSchema>;

export const SalesOrderItemInsertSchema = SalesOrderItemSchema.omit({ id: true, owner: true, creation: true, modified: true, modified_by: true });
export type SalesOrderItemInsert = z.infer<typeof SalesOrderItemInsertSchema>;